+++ lithops/serverless/backends/azure_containers/azure_containers.py
@@ -58,17 +58,16 @@ class AzureContainerAppBackend:
         msg = COMPUTE_CLI_MSG.format('Azure Container Apps')
         logger.info(f"{msg} - Location: {self.location}")
 
-    def _format_containerapp_name(self, runtime_name, runtime_memory=None):
+    def _format_containerapp_name(self, runtime_name, runtime_memory, version=__version__):
         """
         Formates the conatiner app name
         """
-        ver = __version__
         inv_type = self.invocation_type
         ac_name = self.storage_account_name
-        name = f'{ac_name}-{runtime_name}-{ver}-{inv_type}-{runtime_memory}'
+        name = f'{ac_name}-{runtime_name}-{version}-{inv_type}-{runtime_memory}'
         name_hash = hashlib.sha1(name.encode("utf-8")).hexdigest()[:10]
 
-        return f'lithops-ca-runtime-{name_hash}'
+        return f'lithops-runtime-v{version.replace(".", "")}-{name_hash}'
 
     def _get_default_runtime_image_name(self):
         """
@@ -76,7 +75,7 @@ class AzureContainerAppBackend:
         """
         revision = 'latest' if 'dev' in __version__ else __version__
         return utils.get_default_container_name(
-            self.name, self.ac_config, 'lithops-ca-default', revision
+            self.name, self.ac_config, 'lithops-azurecontainers-default', revision
         )
 
     def deploy_runtime(self, runtime_name, memory, timeout):
@@ -204,12 +203,12 @@ class AzureContainerAppBackend:
         finally:
             os.remove(config.CA_JSON_LOCATION)
 
-    def delete_runtime(self, runtime_name, memory):
+    def delete_runtime(self, runtime_name, memory, version=__version__):
         """
         Deletes a runtime
         """
         logger.info(f'Deleting runtime: {runtime_name} - {memory}MB')
-        containerapp_name = self._format_containerapp_name(runtime_name, memory)
+        containerapp_name = self._format_containerapp_name(runtime_name, memory, version)
         cmd = f'az containerapp delete --name {containerapp_name} --resource-group {self.resource_group} -y'
         utils.run_command(cmd)
 
@@ -231,14 +230,14 @@ class AzureContainerAppBackend:
 
             return activation_id
 
-    def get_runtime_key(self, runtime_name, runtime_memory):
+    def get_runtime_key(self, runtime_name, runtime_memory, version=__version__):
         """
         Method that creates and returns the runtime key.
         Runtime keys are used to uniquely identify runtimes within the storage,
         in order to know which runtimes are installed and which not.
         """
-        containerapp_name = self._format_containerapp_name(runtime_name, runtime_memory)
-        runtime_key = os.path.join(self.name, __version__, containerapp_name)
+        containerapp_name = self._format_containerapp_name(runtime_name, runtime_memory, version)
+        runtime_key = os.path.join(self.name, version, containerapp_name)
 
         return runtime_key
 
@@ -251,7 +250,7 @@ class AzureContainerAppBackend:
         runtimes = self.list_runtimes()
 
         for runtime_name, runtime_memory, version in runtimes:
-            self.delete_runtime(runtime_name, runtime_memory)
+            self.delete_runtime(runtime_name, runtime_memory, version)
 
     def _generate_runtime_meta(self, runtime_name, memory):
         """