+++ lithops/serverless/backends/k8s/config.py
@@ -15,7 +15,6 @@
 #
 
 import os
-from lithops.version import __version__
 
 DEFAULT_CONFIG_KEYS = {
     'runtime_timeout': 600,  # Default: 10 minutes