+++ lithops/serverless/backends/openwhisk/openwhisk.py
@@ -51,7 +51,7 @@ class OpenWhiskBackend:
         logger.debug(f"Set OpenWhisk Insecure to {self.insecure}")
 
         self.user_key = self.api_key[:5]
-        self.package = f'lithops_v{__version__}_{self.user_key}'
+        self.package = f'lithops_{self.user_key}'
 
         self.cf_client = OpenWhiskClient(
             endpoint=self.endpoint,
@@ -64,15 +64,15 @@ class OpenWhiskBackend:
         msg = COMPUTE_CLI_MSG.format('OpenWhisk')
         logger.info(f"{msg} - Namespace: {self.namespace}")
 
-    def _format_function_name(self, runtime_name, runtime_memory):
+    def _format_function_name(self, runtime_name, runtime_memory, version=__version__):
         runtime_name = runtime_name.replace('/', '_').replace(':', '_')
-        return f'{runtime_name}_{runtime_memory}MB'
+        return f'{runtime_name}_{runtime_memory}MB_{version}'
 
     def _unformat_function_name(self, action_name):
-        runtime_name, memory = action_name.rsplit('_', 1)
-        image_name = runtime_name.replace('_', '/', 1)
+        runtime_name, memory, version = action_name.rsplit('_', 2)
+        image_name = runtime_name.replace('_', '/', 2)
         image_name = image_name.replace('_', ':', -1)
-        return image_name, int(memory.replace('MB', ''))
+        return version, image_name, int(memory.replace('MB', ''))
 
     def _get_default_runtime_image_name(self):
         try:
@@ -127,12 +127,12 @@ class OpenWhiskBackend:
 
         return self._generate_runtime_meta(docker_image_name, memory)
 
-    def delete_runtime(self, docker_image_name, memory):
+    def delete_runtime(self, docker_image_name, memory, version=__version__):
         """
         Deletes a runtime
         """
         logger.info(f'Deleting runtime: {docker_image_name} - {memory}MB')
-        action_name = self._format_function_name(docker_image_name, memory)
+        action_name = self._format_function_name(docker_image_name, memory, version)
         self.cf_client.delete_action(self.package, action_name)
 
     def clean(self):
@@ -158,14 +158,13 @@ class OpenWhiskBackend:
         runtimes = []
 
         packages = self.cf_client.list_packages()
-        for package in packages:
-            if package['name'].startswith('lithops_v'):
-                version = package['name'].replace('lithops_v', '').split('_')[0]
-                actions = self.cf_client.list_actions(package['name'])
+        for pkg in packages:
+            if pkg['name'] == self.package:
+                actions = self.cf_client.list_actions(pkg['name'])
                 for action in actions:
-                    action_image_name, memory = self._unformat_function_name(action['name'])
-                    if docker_image_name == action_image_name or docker_image_name == 'all':
-                        runtimes.append((action_image_name, memory, version))
+                    version, image_name, memory = self._unformat_function_name(action['name'])
+                    if docker_image_name == image_name or docker_image_name == 'all':
+                        runtimes.append((image_name, memory, version))
         return runtimes
 
     def invoke(self, docker_image_name, runtime_memory, payload):
@@ -179,14 +178,14 @@ class OpenWhiskBackend:
 
         return activation_id
 
-    def get_runtime_key(self, docker_image_name, runtime_memory):
+    def get_runtime_key(self, docker_image_name, runtime_memory, version=__version__):
         """
         Method that creates and returns the runtime key.
         Runtime keys are used to uniquely identify runtimes within the storage,
         in order to know which runtimes are installed and which not.
         """
-        action_name = self._format_function_name(docker_image_name, runtime_memory)
-        runtime_key = os.path.join(self.name, __version__, self.namespace, action_name)
+        action_name = self._format_function_name(docker_image_name, runtime_memory, version)
+        runtime_key = os.path.join(self.name, version, self.namespace, action_name)
 
         return runtime_key